+++ internal/command/put.go
@@ -291,7 +291,7 @@ func getOutputWriter(cmd *cobra.Command, in io.Writer, file string, out string) 
 
 func putCommand() *cobra.Command {
 	var fileFlag, selectorFlag, parserFlag, readParserFlag, writeParserFlag, outFlag, valueFlag string
-	var multiFlag, compactFlag, mergeInputDocumentsFlag bool
+	var multiFlag, compactFlag, mergeInputDocumentsFlag, escapeHTMLFlag bool
 
 	cmd := &cobra.Command{
 		Use:   "put -f <file> -s <selector>",
@@ -316,6 +316,7 @@ func putCommand() *cobra.Command {
 	cmd.PersistentFlags().BoolVarP(&compactFlag, "compact", "c", false, "Compact the output by removing all pretty-printing where possible.")
 	cmd.PersistentFlags().BoolVar(&mergeInputDocumentsFlag, "merge-input-documents", false, "Merge multiple input documents into an array.")
 	cmd.PersistentFlags().StringVarP(&valueFlag, "value", "v", "", "Value to put.")
+	cmd.PersistentFlags().BoolVar(&escapeHTMLFlag, "escape-html", true, "Escape HTML tags when writing output.")
 
 	_ = cmd.MarkPersistentFlagFilename("file")
 
@@ -337,6 +338,7 @@ type genericPutOptions struct {
 	Multi               bool
 	Compact             bool
 	MergeInputDocuments bool
+	EscapeHTML          bool
 }
 
 func getGenericInit(cmd *cobra.Command, args []string) func(options genericPutOptions) genericPutOptions {
@@ -351,6 +353,7 @@ func getGenericInit(cmd *cobra.Command, args []string) func(options genericPutOp
 		opts.Compact, _ = cmd.Flags().GetBool("compact")
 		opts.MergeInputDocuments, _ = cmd.Flags().GetBool("merge-input-documents")
 		opts.Value, _ = cmd.Flags().GetString("value")
+		opts.EscapeHTML, _ = cmd.Flags().GetBool("escape-html")
 
 		if opts.Selector == "" && len(args) > 0 {
 			opts.Selector = args[0]
@@ -404,7 +407,9 @@ func runGenericPutCommand(opts genericPutOptions, cmd *cobra.Command) error {
 		return err
 	}
 
-	writeOptions := make([]storage.ReadWriteOption, 0)
+	writeOptions := []storage.ReadWriteOption{
+		storage.EscapeHTMLOption(opts.EscapeHTML),
+	}
 
 	if opts.Compact {
 		writeOptions = append(writeOptions, storage.PrettyPrintOption(false))